chore: remove src2
