--- src2/builder.ts
@@ -1,3 +1,0 @@
-export class MonorepoNetworker {
-  private sides: string[] = [];
-}
