+++ src3/side.ts
@@ -1,0 +1,7 @@
+export interface AcceptedEvents {
+  [eventName: string]: (...args: any[]) => any;
+}
+
+export class NetworkSide<T extends AcceptedEvents> {
+  constructor(public readonly name: string) {}
+}
